Moved where dependencies are specified
