+++ dlhub_sdk/models/servables/__init__.py
@@ -1,9 +1,3 @@
-import pkg_resources
-import importlib
-import requests
-import os
-
-
 from dlhub_sdk.models import BaseMetadataModel
 
 
@@ -27,8 +21,7 @@ class BaseServableModel(BaseMetadataModel):
         self._output['servable'] = {
             'methods': {'run': {}},
             'shim': self._get_handler(),
-            'type': self._get_type(),
-            'dependencies': {'python': {}}
+            'type': self._get_type()
         }
 
     def _get_handler(self):
@@ -85,70 +78,3 @@ class BaseServableModel(BaseMetadataModel):
             raise ValueError('Outputs have not been defined')
 
         return super(BaseServableModel, self).to_dict(simplify_paths)
-
-    def add_requirement(self, library, version=None):
-        """Add a required Python library.
-
-        The name of the library should be either the name on PyPI, or a URL for the git repository
-        holding the code (e.g., ``git+https://github.com/DLHub-Argonne/dlhub_sdk.git``)
-
-        Args:
-            library (string): Name of library
-            version (string): Required version. 'latest' to use the most recent version on PyPi (if
-                available). 'detect' will attempt to find the version of the library installed on
-                the computer running this software. Default is ``None``
-        """
-
-        # Attempt to determine the version automatically
-        if version == "detect":
-            try:
-                module = importlib.import_module(library)
-                version = module.__version__
-            except Exception:
-                version = pkg_resources.get_distribution(library).version
-        elif version == "latest":
-            pypi_req = requests.get('https://pypi.org/pypi/{}/json'.format(library))
-            version = pypi_req.json()['info']['version']
-
-        # Set the requirements
-        self._output["servable"]["dependencies"]["python"][library] = version
-        return self
-
-    def add_requirements(self, requirements):
-        """Add a dictionary of requirements
-
-        Utility wrapper for `add_requirement`
-
-        Args:
-            requirements (dict): Keys are names of library (str), values are the version
-        """
-        for p, v in requirements.items():
-            self.add_requirement(p, v)
-        return self
-
-    def parse_repo2docker_configuration(self, directory=None):
-        """Gathers information about required environment from repo2docker configuration files.
-
-        See https://repo2docker.readthedocs.io/en/latest/config_files.html for more details
-
-        Args:
-            directory (str): Path to directory containing configuration files
-                (default: current working directory)
-        """
-
-        # Get a list of all files
-        config_files = ['environment.yml', 'requirements.txt', 'setup.py', 'REQUIRE', 'install.R',
-                        'apt.txt', 'DESCRIPTION', 'manifest.xml', 'postBuild', 'start',
-                        'runtime.txt', 'default.nix', 'Dockerfile']
-
-        # Get the directory name if `None`
-        if directory is None:
-            directory = os.getcwd()
-
-        # Add every file we can find
-        for file in config_files:
-            path = os.path.join(directory, file)
-            if os.path.isfile(path):
-                self.add_file(path)
-
-        return self
